requirements.txt
fastapi==0.118.0
matplotlib==3.10.6
matplotlib-inline==0.1.7
numpy==2.1.2
opencv-python==4.12.0.88
pandas==2.3.2
pillow==11.0.0
streamlit==1.50.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
ultralytics==8.3.203
ultralytics-thop==2.0.17
uvicorn==0.37.0